+++ tests/bootstrap.php
@@ -15,9 +15,27 @@ namespace GuzzleHttp\Test {
     });
 }
 
-// Override curl_setopt_array() and curl_multi_setopt() to get the last set curl options
+// Override curl_setopt(), curl_setopt_array(), and curl_multi_setopt() to get the last set curl options
 
 namespace GuzzleHttp\Handler {
+    function curl_setopt($handle, int $option, $value)
+    {
+        if (!empty($_SERVER['curl_test'])) {
+            if ($option === \CURLOPT_CUSTOMREQUEST) {
+                $_SERVER['_curl'] = [];
+            }
+            $_SERVER['_curl'][$option] = $value;
+        } else {
+            unset($_SERVER['_curl']);
+        }
+
+        if (isset($_SERVER['curl_setopt_fail']) && (int) $_SERVER['curl_setopt_fail'] === $option) {
+            return false;
+        }
+
+        return \curl_setopt($handle, $option, $value);
+    }
+
     function curl_setopt_array($handle, array $options)
     {
         if (!empty($_SERVER['curl_test'])) {